+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
- - Avoid some over-generalizations by only whistling on terms with the same redex operator ([issue #9](https://github.com/mazeppa-dev/mazeppa/issues/9)).
+ - Avoid some over-generalizations by checking redex signatures (issues https://github.com/mazeppa-dev/mazeppa/issues/9, https://github.com/mazeppa-dev/mazeppa/issues/11).
 
 ## 0.1.2 - 2024-07-20